moved duplicates to make deploying easier
