+++ ChartPackage/main/default/lwc/utilsPrivate/__tests__/normalizeAriaAttribute.spec.js
