+++ ChartPackage/main/default/lwc/utilsPrivate/__tests__/normalizeString.spec.js
