+++ ChartPackage/main/default/lwc/utilsPrivate/guid.js
@@ -1,8 +1,15 @@
+/*
+ * Copyright (c) 2019, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
 export function guid() {
     function s4() {
         return Math.floor((1 + Math.random()) * 0x10000)
             .toString(16)
-            .substring(1)
+            .substring(1);
     }
 
     return (
@@ -18,5 +25,5 @@ export function guid() {
         s4() +
         s4() +
         s4()
-    )
+    );
 }
