+++ ChartPackage/main/default/lwc/utilsPrivate/keyboard.js
@@ -1,3 +1,10 @@
+/*
+ * Copyright (c) 2019, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
 export const keyCodes = {
     tab: 9,
     backspace: 8,
@@ -14,4 +21,55 @@ export const keyCodes = {
     down: 40,
     delete: 46,
     shift: 16
+};
+
+export function normalizeKeyValue(value) {
+    switch (value) {
+        case 'Spacebar':
+            return ' ';
+        case 'Esc':
+            return 'Escape';
+        case 'Del':
+            return 'Delete';
+        case 'Left':
+            return 'ArrowLeft';
+        case 'Right':
+            return 'ArrowRight';
+        case 'Down':
+            return 'ArrowDown';
+        case 'Up':
+            return 'ArrowUp';
+        default:
+            return value;
+    }
+}
+
+const buffer = {};
+
+export function isShiftMetaOrControlKey(event) {
+    return event.shiftKey || event.metaKey || event.ctrlKey;
+}
+
+export function runActionOnBufferedTypedCharacters(event, action) {
+    const letter = event.key;
+
+    if (letter.length > 1) {
+        return;
+    }
+
+    if (buffer._clearBufferId) {
+        clearTimeout(buffer._clearBufferId);
+    }
+
+    buffer._keyBuffer = buffer._keyBuffer || [];
+    buffer._keyBuffer.push(letter);
+
+    const matchText = buffer._keyBuffer.join('').toLowerCase();
+
+    action(matchText);
+
+    // eslint-disable-next-line @lwc/lwc/no-async-operation
+    buffer._clearBufferId = setTimeout(() => {
+        buffer._keyBuffer = [];
+    }, 700);
 }
