+++ ChartPackage/main/default/lwc/utilsPrivate/utilsPrivate.js-meta.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>52.0</apiVersion>
-    <isExposed>false</isExposed>
+
+<LightningComponentBundle xmlns="xmlns=http://soap.sforce.com/2006/04/metadata">
+    
+  <isExposed>false</isExposed>
+    
+  <apiVersion>50.0</apiVersion>
+    
 </LightningComponentBundle>
