--- TitanHubPackage/main/default/lwc/utilsPrivate/utilsPrivate.js-meta.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<LightningComponentBundle xmlns="xmlns=http://soap.sforce.com/2006/04/metadata">
-    
-  <isExposed>false</isExposed>
-    
-  <apiVersion>50.0</apiVersion>
-    
-</LightningComponentBundle>
