--- TrialoftheTitansPackage/main/default/lwc/radar/radar.html
@@ -1,6 +1,0 @@
-<template>
-    <div lwc:dom="manual" class="d3 slds-align_absolute-center"></div>
-    
-    <!-- Rerenders component on data refresh  -->
-    <div hidden>{dat}</div>
-</template>
