--- TrialoftheTitansPackage/main/default/lwc/radarCarousel/radarCarousel.html
@@ -1,39 +1,0 @@
-<template>
-    <div class="slds-carousel">
-        <div class="slds-carousel__stage">
-            <template if:false={disableAutoScroll}>
-                <span class="slds-carousel__autoplay">
-                    <button class="slds-button slds-button_icon slds-button_icon-border-filled slds-button_icon-x-small"
-                        aria-pressed={ariaPressed} title={togglePlayString} onclick={toggleAutoScroll}>
-                        <svg class="slds-button__icon" aria-hidden="true">
-                            <use xlink:href={svgHref}></use>
-                        </svg>
-                        <span class="slds-assistive-text">{togglePlayString}</span>
-                    </button>
-                </span>
-            </template>
-
-            <div class="slds-carousel__panels" style={carouselPanelsStyle} ontouchstart={handleTouchStart}
-                ontouchend={handleTouchEnd}>
-                <template for:each={avgScores} for:item="avgScore" for:index="avgIndex">
-                    <c-radar-carousel-image href="#" src="" header={avgScore.category} scores={avgScore.axes}
-                        key={avgScore.key} data-index={avgIndex} onprivateimageregister={handlePrivateImageRegister}>
-                    </c-radar-carousel-image>
-                </template>
-            </div>
-
-            <ul class="slds-carousel__indicators" role="tablist" onkeydown={keyDownHandler}>
-                <template for:each={paginationItems} for:item="paginationItem" for:index="paginationIndex">
-                    <li class="slds-carousel__indicator" role="presentation" onclick={onItemSelect}
-                        key={paginationItem.key} data-index={paginationIndex}>
-                        <a id={paginationItem.id} class={paginationItem.className} href="#" role="tab"
-                            tabindex={paginationItem.tabIndex} aria-selected={paginationItem.ariaSelected}
-                            title={paginationItem.tabTitle}>
-                            <span class="slds-assistive-text">{paginationItem.tabTitle}</span>
-                        </a>
-                    </li>
-                </template>
-            </ul>
-        </div>
-    </div>
-</template>
