--- TrialoftheTitansPackage/main/default/lwc/radarCarouselImage/radarCarouselImage.html
@@ -1,13 +1,0 @@
-<template>
-    <div id="carousel-image" role="tabpanel" aria-hidden={ariaHidden}>
-        <a href={href} class="slds-carousel__panel-action slds-text-link_reset" tabindex={tabIndex}>
-            <div class="slds-carousel__image">
-                <c-radar dat={scores}></c-radar>
-            </div>
-            <div class="slds-carousel__content">
-                <h2 class="slds-carousel__content-title">{header}</h2>
-                <p>{description}</p>
-            </div>
-        </a>
-    </div>
-</template>
