--- TrialoftheTitansPackage/main/default/lwc/utilsPrivate/assert.js
@@ -1,1 +1,0 @@
-export function assert(condition, message) {}
