--- TrialoftheTitansPackage/main/default/lwc/utilsPrivate/guid.js
@@ -1,22 +1,0 @@
-export function guid() {
-    function s4() {
-        return Math.floor((1 + Math.random()) * 0x10000)
-            .toString(16)
-            .substring(1)
-    }
-
-    return (
-        s4() +
-        s4() +
-        '-' +
-        s4() +
-        '-' +
-        s4() +
-        '-' +
-        s4() +
-        '-' +
-        s4() +
-        s4() +
-        s4()
-    )
-}
