--- TrialoftheTitansPackage/main/default/lwc/utilsPrivate/keyboard.js
@@ -1,17 +1,0 @@
-export const keyCodes = {
-    tab: 9,
-    backspace: 8,
-    enter: 13,
-    escape: 27,
-    space: 32,
-    pageup: 33,
-    pagedown: 34,
-    end: 35,
-    home: 36,
-    left: 37,
-    up: 38,
-    right: 39,
-    down: 40,
-    delete: 46,
-    shift: 16
-}
