--- TrialoftheTitansPackage/main/default/lwc/utilsPrivate/normalize.js
@@ -1,3 +1,0 @@
-export function normalizeBoolean(value) {
-    return typeof value === 'string' || !!value;
-}
