--- TrialoftheTitansPackage/main/default/lwc/utilsPrivate/utilsPrivate.js
@@ -1,9 +1,0 @@
-import { LightningElement } from 'lwc';
-
-export { assert } from './assert'
-
-export { guid } from './guid'
-
-export { normalizeBoolean } from './normalize'
-
-export { keyCodes } from './keyboard'
