--- TrialoftheTitansPackage/main/default/staticresources/RadarImage.resource-meta.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<StaticResource xmlns="http://soap.sforce.com/2006/04/metadata">
-    <cacheControl>Public</cacheControl>
-    <contentType>application/x-zip-compressed</contentType>
-</StaticResource>
